removed some unused code, moved declarations to headers, renaming of some methods

+++ src/st-machine.h
@@ -12,12 +12,12 @@
 #include <setjmp.h>
 #include "st-types.h"
 
-/* cache size must be a power of 2 */
+// cache size must be a power of 2
 #define ST_METHOD_CACHE_SIZE      1024
 #define ST_METHOD_CACHE_MASK      (ST_METHOD_CACHE_SIZE - 1)
 #define ST_METHOD_CACHE_HASH(k, s) ((k) ^ (s))
 
-#define ST_NUM_GLOBALS 36
+#define ST_NUM_GLOBALS   36
 #define ST_NUM_SELECTORS 24
 
 typedef struct st_method_cache {
@@ -38,21 +38,19 @@ struct st_machine {
 	st_oop lookup_class;
 	st_oop message_receiver;
 	st_oop message_selector;
-	int message_argcount;
+	size_t message_argcount;
 	st_oop new_method;
 	bool success;
 	st_uint ip;
 	st_uint sp;
 	jmp_buf main_loop;
-	
 	st_method_cache method_cache[ST_METHOD_CACHE_SIZE];
-	
 	st_oop globals[ST_NUM_GLOBALS];
 	st_oop selectors[ST_NUM_SELECTORS];
-	
 };
 
 extern st_machine __machine;
+
 #define ST_STACK_POP(machine)          (machine->stack[--machine->sp])
 #define ST_STACK_PUSH(machine, oop)    (machine->stack[machine->sp++] = oop)
 #define ST_STACK_PEEK(machine)         (machine->stack[machine->sp-1])
